v0.4.142: split AssignmentHelpers.swift into 8 focused files (#442)
Pure relocation; no behaviour changes. swift test green pre- and
post-split.
